Issue #1792 Moved quoting methods to ComposeQuote Class
